created bzero, memcpy, memccpy, memdel, memset, strclr, strdel, striter, striteri, strnew. Staging a.out for debug (memccpy and memcpy)

+++ ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: fpolyans <fpolyans@42.us.org>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/22 00:57:13 by fpolyans          #+#    #+#             */
-/*   Updated: 2017/09/28 18:48:34 by fpolyans         ###   ########.fr       */
+/*   Updated: 2017/10/01 13:46:19 by fpolyans         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,5 +31,5 @@ char	*ft_strtrim(char const *s)
 		}
 		i++;
 	}
-	return (ft_stricut(s, (size_t)start, (size_t)end));
+	return (ft_stricut(s, start, end));
 }
